requirements.txt
# API-Based Generation (Recommended - No GPU needed!)
huggingface-hub>=0.20.0
replicate>=0.20.0  # Optional: $10 free credits

# Core Dependencies (for local generation)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
omegaconf>=2.3.0

# Image/Video Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3

# Model Enhancement
xformers>=0.0.22  # Optional but recommended for speed
einops>=0.7.0
ftfy>=6.1.1

# Upscaling & Enhancement
realesrgan>=0.3.0
basicsr>=1.4.2

# Audio (Optional TTS)
TTS>=0.20.0

# Web Interface
gradio>=4.0.0
flask>=3.0.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.66.0
psutil>=5.9.0
pyyaml>=6.0.0
requests>=2.31.0
huggingface-hub>=0.19.0

# Development
pytest>=7.4.0
black>=23.0.0
